Refactor: consolidate encoding utilities (base64/hex/ArrayBuffer)
